move code from 'lib/' to 'src/'
