+++ .eslintignore
@@ -1,1 +1,1 @@
-lib/Grammar.js
+src/Grammar.js
