+++ src/DigestAuthentication.js
