+++ src/Exceptions.d.ts
