+++ src/Registrator.d.ts
