+++ src/Registrator.js
