+++ src/Socket.d.ts
