+++ src/URI.d.ts
