+++ src/WebSocketInterface.d.ts
