+++ src/sanityCheck.js
